pyproject.toml
[tool.poetry]
name = "root"
version = "0.1.0"
description = ""
authors = ["nurali <nuralirahamatov323@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-mptt = "^0.14.0"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
